requirements.txt
torch==2.0.1
transformers==4.28.1
python-dotenv == 0.21.0
requests == 2.28.1
uvicorn[standard]==0.22.0
fastapi==0.95.1
pydantic==1.10.12